Chat Server/requirements.txt
langgraph
langchain
langchain-mcp-adapters
fastmcp
fastapi
uvicorn
langchain-google-genai
python-dotenv